+++ tests/Unit/SqlQueryTest.php
@@ -2,7 +2,6 @@
 
 namespace Tests\Unit;
 
-use Illuminate\Support\Carbon;
 use Onlime\LaravelSqlReporter\SqlQuery;
 
 class SqlQueryTest extends UnitTestCase